+++ packages/q3-admin/src/components/list.jsx
@@ -35,6 +35,7 @@ const List = ({ children, ...rest }) => {
     resourceNameSingular,
     collectionName,
     location,
+    url,
     ...state
   } = React.useContext(Context);
 
@@ -64,7 +65,7 @@ const List = ({ children, ...rest }) => {
     if (state.fetchingError) return <ErrorView />;
     if (!rows.length) return <EmptyView />;
     return (
-      <Table {...state} {...rest}>
+      <Table {...state} {...rest} id={url}>
         {children(rows)}
       </Table>
     );