+++ packages/q3-ui/CHANGELOG.md
@@ -3,10 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [1.7.3](https://github.com/3merge/q/compare/v1.7.2...v1.7.3) (2020-01-27)
+# [1.8.0](https://github.com/3merge/q/compare/v1.7.2...v1.8.0) (2020-01-23)
 
-**Note:** Version bump only for package q3-ui
 
+### Features
+
+* sorting and pagination ([c4a35e8](https://github.com/3merge/q/commit/c4a35e8225f474449df6846d0561f5c34f87ceae))